% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_errorcrosses.R
\name{plot_errorcrosses}
\alias{plot_errorcrosses}
\title{Plot pairwise with two variables}
\usage{
plot_errorcrosses(
  datasetx = data_color,
  datasety = data_control,
  traitx = "red",
  traity = "Nb_eggs",
  xaxis_labelprint = "Color: Red",
  yaxis_labelprint = "Number of eggs"
)
}
\arguments{
\item{xaxis_labelprint}{label of the x axis}

\item{yaxis_labelprint}{label of the y axis}

\item{dataset1}{name of the first dataset}

\item{dataset2}{name of the second dataset}

\item{trait1}{variable from the first dataset}

\item{trait2}{variable from the second dataset}
}
\value{
plot
}
\description{
Create a bivariate plot with two variables
}
\examples{
plot_realdata(datasetx = data_color,datasety = data_control,traitx="red",traity="Nb_eggs",xaxis_labelprint = "Color: Red",yaxis_labelprint = "Number of eggs") 
}
